feat: Added `@unpack_option` macro
feat: Added `UnionExpr` expression type

refactor: Renamed `StructDefField` -> `TypedVar` type, added an alias for `StructDefField`

test: Reorganized test files
