+++ src/parse_kwargs/kv_expr_parser.jl
@@ -6,9 +6,8 @@ end
 
 Base.:(==)(x::KVSpecExpr, y::KVSpecExpr) =  all(getfield(x,k) == getfield(y,k) for k in fieldnames(KVSpecExpr))
 
-
 function parse_kv_spec_from_type_expr(expr)
-    f = _from_expr(StructDefField, expr)
+    f = _from_expr(TypedVar, expr)
     f isa Exception && return f 
     key = f.name 
     expected_types = Set()
