+++ src/parsing/_parsing.jl
@@ -1,8 +1,8 @@
 include("util.jl")
 include("types.jl")
 include("expr_types/_expr_types.jl")
+include("macrocall.jl")
 include("parsing_macros.jl")
 include("parsing.jl")
-include("macrocall.jl")
 include("funcs.jl")
 include("struct.jl")
