+++ src/parsing/expr_types/_expr_types.jl
@@ -1,3 +1,4 @@
 include("base.jl")
 include("composite.jl")
+include("typed_var.jl")
 include("syntax_sugar.jl")
