+++ test/expressions/_test_expressions.jl
@@ -1,0 +1,4 @@
+include("test_exprs.jl")
+include("test_functions.jl")
+include("test_struct.jl")
+include("test_macros.jl")
