+++ test/runtests.jl
@@ -1,1 +1,5 @@
-include("TestMacroUtilities.jl")
+using Test 
+
+@testset "MacroUtilities.jl" begin 
+    include("TestMacroUtilities.jl")
+end
